requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
pycodestyle==2.6.0
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
uuid==1.30
whitenoise==5.2.0
